folder arrangement and start design of the result component
